reporting/crash-reporting-017.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Crash report includes crash_report_api dictionary</title>
<link rel="help" href="https://wicg.github.io/crash-reporting/">
<meta name="assert" content="The crash report body must include a `crash_report_api` JSON dictionary containing the deserialized key/value data from the user agent's crash report buffer.">
<script>
const params = new URLSearchParams(window.location.search);
const isPopup = params.has('popup');

if (!isPopup) {
    document.write('<script src="/resources/testharness.js"><\/script>');
    document.write('<script src="/resources/testharnessreport.js"><\/script>');
}
</script>
<script src="/common/utils.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script>
if (isPopup) {
    (async function() {
        try {
            // 1. Call initialize(1024) and set('key', 'value').
            if (window.crashReport) {
                await window.crashReport.initialize(1024);
                window.crashReport.set('key', 'value');
            }
        } catch (e) {
            // Ignore errors
        }

        // 2. Trigger a renderer crash.
        // We force an Out-Of-Memory (OOM) crash which is a valid crash reason in the spec.
        let a = [];
        while (true) {
            a.push(new Array(1000000).join('a'));
        }
    })();
} else {
    promise_test(async t => {
        const hostInfo = get_host_info();
        const mainOrigin = hostInfo.HTTPS_ORIGIN;
        // Use a remote origin for the popup to ensure Site Isolation places it in a separate 
        // renderer process, preventing the main test window from crashing.
        const remoteOrigin = hostInfo.HTTPS_REMOTE_ORIGIN;

        const reportID = token();
        const endpointUrl = `${mainOrigin}/reporting/resources/report.py?op=put&reportID=${reportID}`;
        const retrieveUrl = `${mainOrigin}/reporting/resources/report.py?op=retrieve_report&timeout=5&reportID=${reportID}`;

        // Pre-condition: Reporting-Endpoints configured.
        // We inject the header into the popup response using WPT server pipes.
        const pipeVal = `header(Reporting-Endpoints,default="${endpointUrl}")`;
        const popupUrl = `${remoteOrigin}${window.location.pathname}?popup=1&pipe=${encodeURIComponent(pipeVal)}`;

        // Open with noopener to completely detach the popup. This prevents its intentional crash
        // from being detected by the WPT runner (via window.opener links) and marking the test as CRASH.
        const popup = window.open(popupUrl, '_blank', 'noopener');
        t.add_cleanup(() => {
            if (popup) {
                try { popup.close(); } catch (e) {}
            }
        });

        // 3. Inspect the delivered crash report.
        let report = null;
        // Poll the WPT reporting stash for up to 15 seconds
        for (let i = 0; i < 15; i++) {
            await new Promise(r => step_timeout(r, 1000));
            const res = await fetch(retrieveUrl);
            if (res.ok) {
                const data = await res.json();
                if (data && data.length > 0) {
                    report = data[0];
                    break;
                }
            }
        }

        assert_true(report !== null, "Crash report should be delivered to the endpoint.");
        assert_equals(report.type, "crash", "Report type should be 'crash'.");
        assert_true("body" in report, "Report should have a body.");
        
        // Expected Result: The crash report body contains a 'crash_report_api' dictionary with {"key": "value"}.
        assert_true("crash_report_api" in report.body, "Report body should contain 'crash_report_api'.");
        assert_equals(typeof report.body.crash_report_api, "object", "crash_report_api should be a JSON dictionary.");
        assert_equals(report.body.crash_report_api.key, "value", "crash_report_api should contain the deserialized key/value data.");

    }, "Crash report includes crash_report_api dictionary");
}
</script>
